backend/requirements.txt
Flask
flask_pymongo
flask_jwt_extended
pyjwt
google-generativeai
python-dotenv
flask_socketio
eventlet
flask_cors
